requirements.txt
faker==25.9.1
Flask==3.0.3
pymongo==4.10.1
python-dotenv==1.0.1
voyageai==0.3.5
